Refactor and disable appointment management features
This commit includes significant changes to the SmartClinic application, primarily focusing on commenting out existing code in various controllers and services related to appointment and doctor management. Key controllers such as `AppointmentsController`, `DoctorsController`, and `PatientsController` have had their action methods disabled.

Additionally, the project file has been updated to include a new uploads folder, and the connection string in `appsettings.json` has been modified to point to a different database. Several binary image files have been deleted.

Changes to pagination logic in the `Pagination` and `PagingParams` classes have been made, along with updates to the `DoctorMappingExtensions` class to adjust how doctor data is mapped to response DTOs. New classes and interfaces, including `IDto`, `GetAllDoctorsParams`, and `DoctorSpecifications`, have been added to enhance the application's structure and functionality.

+++ SmartClinic.Application/Features/Doctors/Query/DTOs/GetDoctorWithAvailableAppointment/GetDoctorWithAvailableAppointment.cs
@@ -4,7 +4,7 @@ public record GetDoctorWithAvailableAppointment(
          string? LastName,
          string UserEmail,
          string UserPhoneNumber,
-        byte Age,
+        int Age,
         string Address,
         string? Description,
         int? WaitingTime,
